Give resource cost as double to statistics.
Previously resource cost was given as unsigned, which implied fractions
had to be rounded as well. This meant loss of resolution for DoTs like
Rend and Consecration wherein the resource cost for a given damage
report is a fraction of the actual resource cost.

+++ Character/Class/Paladin/Spells/Consecration.cpp
@@ -54,7 +54,7 @@ void Consecration::calculate_damage() {
     const double damage_dealt = ((full_duration_dmg + bonus_from_spell_dmg) / 4 + tick_rest) * resist_mod * pchar->get_stats()->get_spell_dmg_mod(MagicSchool::Holy);
     tick_rest = tick_rest + damage_dealt - round(damage_dealt);
 
-    add_hit_dmg(static_cast<int>(round(damage_dealt)), resource_cost / 4, pchar->global_cooldown() / 4);
+    add_hit_dmg(static_cast<int>(round(damage_dealt)), static_cast<double>(resource_cost) / 4, pchar->global_cooldown() / 4);
 }
 
 void Consecration::spell_effect() {
